AnalyseAgent/requirements.txt
mcp
strands-agents
strands-agents-tools
a2a-sdk
uvicorn
fastapi
anthropic
ollama
requests
